Revert "Rename legacy modules"
