+++ src/legacy/modules/products.js
